Split structure from actions on the domain.
